App 정리 및 my page, 회원가입 (#11)
* 게시판과 회원정보를 나눔, 마이페이지 및 회원정보부분 추가
