+++ static/fonts/fontawesome-webfont.woff2
