+++ templates/base.html
@@ -26,7 +26,6 @@
     <![endif]-->
 </head>
 <body>
-
     {% include 'header.html' %}
 
     {% block content %}
