+++ templates/profile/edit_info.html
@@ -1,0 +1,41 @@
+{% extends 'base.html' %}
+{% block content %}
+    <article class="container">
+        <div class="page-header">
+            <h2>마이 페이지 수정
+                <small>My page edit</small>
+            </h2>
+        </div>
+
+        <div class="col-md-6 col-md-offset-3">
+            {{ err }}
+            <form role="form" method="POST">
+                {% csrf_token %}
+                <label>아이디 : </label>{{ user.username }}<br>
+                <label>이메일 : </label>{{ user.email }}<br>
+
+                <div class="form-group">
+                    <label for="{{ form.name.id_for_label }}">{{ form.name.label }} *</label>
+                    {{ form.name }}
+                </div>
+
+                <label>학번 : </label>{{ user_info.user_num }}<br>
+                <div class="form-group">
+                    <label for="{{ form.phone.id_for_label }}">{{ form.phone.label }} *</label>
+                    {{ form.phone }}
+                </div>
+                <div class="form-group">
+                    <label for="{{ form.is_attend.id_for_label }}">{{ form.is_attend.label }} *</label>
+                    {{ form.is_attend }}
+                </div>
+                <label>가입일자 : </label>{{ user_info.signup_date }}<br>
+
+                <button type="submit" class="btn btn-info">회원정보 수정<i class="fa fa-check spaceLeft"></i></button>
+
+                <a href="{% url 'my_page' %}">
+                    <button type="button" class="btn btn-warning">취소<i class="fa fa-times spaceLeft"></i></button>
+                </a>
+            </form>
+        </div>
+    </article>
+{% endblock %}
