+++ templates/profile/user_active.html
@@ -1,0 +1,4 @@
+{% extends 'base.html' %}
+{% block content %}
+
+{% endblock %}
